\name{RcmdrepackPlugin-internal}
\title{Internal RcmdrepackPlugin objects}
\alias{Arima}
\alias{ArimaMod}
\alias{newHistPrice}
\alias{bc2}
\alias{pow2}
\alias{histprice2}
\alias{decom5}
\alias{forerr}
\alias{predar3}
\alias{acfMod}
\alias{bcMod}
\alias{decomMod}
\alias{foreMod}
\alias{garchMod}
\alias{listArimaModels}
\alias{listGARCHModels}
\alias{opMod}
\alias{pacfMod}
\alias{predARModel}
\alias{selectActiveARModel}
\alias{specMod}
\alias{tsConv}


\description{Internal RcmdrepackPlugin objects.}
\details{These are not to be called by the user.}
\keyword{internal}
