\name{RcmdrepackPlugin-package}
\alias{RcmdrepackPlugin-package}
\alias{RcmdrepackPlugin}
\docType{package}
\title{
Install the Time series  Rcmdr Plug-In
}
\description{
This package provides an Rcmdr "plug-in" for teaching time series.
}
\details{
\tabular{ll}{
Package: \tab RcmdrepackPlugin\cr
Type: \tab Package\cr
Version: \tab 1.0-1\cr
Date: \tab 2007-11-24\cr
License: \tab GPL version 2 or newer\cr
}
}

\author{
Erin Hodgess <hodgesse@uhd.edu>

Maintainer: Erin Hodgess <hodgesse@uhd.edu>
}

\keyword{ package }
\seealso{
\code{\link[Rcmdr]{Rcmdr}}.
}

